+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 - fixed hide error object of type QProcess on WorkProcess class #93
+- fixed settings dhcp for allow to change dhcp configuration
 
 ## [Released]